requirements.txt
beautifulsoup4==4.9.0
psutil==5.7.0
elasticsearch7==7.6.0
requests==2.23.0
elasticsearch==7.0.0
python-dateutil==2.6.1